lock all cgi scripts during a running firmware update so that no other job can interfere.

+++ buildroot-external/package/recovery-system/external/overlay/www/cgi-bin/firmware_update.cgi
@@ -2,6 +2,11 @@
 
 echo -ne "Content-Type: text/html; charset=iso-8859-1\r\n\r\n"
 
+if [ -f /tmp/.runningFirmwareUpdate ]; then
+  echo "ERROR: firmware update is already running"
+  exit 1
+fi
+
 # cleanup first
 rm -f /usr/local/.firmwareUpdate
